rag_workflow/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
faiss-cpu
sentence-transformers
google-genai
